% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_generate.R
\name{project_generate}
\alias{project_generate}
\title{Generate a project folder}
\usage{
project_generate(project_name, category_path = ".", priority = "normal")
}
\arguments{
\item{project_name}{the name of your project. E.g. \code{graph-db-buildout}.
No spaces please -- the text will be used to create a directory path.}

\item{category_path}{path to the category to create the project under
e.g. \verb{personal_work/}. Default is ".", and it is created in reference
to}

\item{priority}{priority level of the project, choose from "someday-maybe",
"normal", "high", or "maximum-effort"}
}
\description{
Creates a project folder in specified category (default is current directory)
includes creation of \code{worklog.md}, and populates the metadata within.
}
